🤖 Agentic AI/class_02/myName/pyproject.toml
[project]
name = "myname"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ArAhmadRaza", email = "arahmadraza5570@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "google-generativeai[gemini-1.5-flash]>=0.8.5",
]

[project.scripts]
myname = "myname:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
